refactor(game): split scene3d.ts 1577 baris → scene3d/ 17 modul domain (zero behavior change)
