+++ apps/game/src/renderer/scene3d/rng.ts
@@ -1,0 +1,18 @@
+// Copyright 2026 GSF-001
+//
+// Licensed under the ARCLUX MMO License v1 (GSF-001) — Source-available, No Commercial Game Clone.
+// See LICENSE-MMO in the repo root. SPDX: LicenseRef-ARCLUX-MMO.
+//
+// scene3d/rng.ts — deterministic RNG (client-side, sama deterministik server
+// §2.3). Moved verbatim dari scene3d.ts — pure utility, zero behavior change.
+
+/** mulberry32 — seeded PRNG. Seed per domain biar layout deterministik. */
+export function mulberry32(seed: number): () => number {
+  let a = seed >>> 0;
+  return () => {
+    a |= 0; a = (a + 0x6d2b79f5) | 0;
+    let t = Math.imul(a ^ (a >>> 15), 1 | a);
+    t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t;
+    return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
+  };
+}
